fix: chain-id fields generated as float32 cannot represent Aurora's chain id
A float32 has a 24-bit mantissa, so integers above 2^24 lose precision:
Aurora (1313161554) silently rounds to 1313161600, corrupting the EIP-712
signing domain and the encoded escrow DstChainId, and making Aurora
impossible to pass through parameter validation.

The codegen pipeline now rewrites chain-id parameters and properties
typed as "number" to "integer" before type generation, so all chain-id
fields in fusionplus and tokens are generated as int. Handwritten
*Fixed mirror types, the escrow extension, and validation follow suit.
The unused CheckChainIdFloat32 validators are removed.

+++ sdk-clients/fusionplus/escrowextension.go
@@ -16,7 +16,7 @@ import (
 type EscrowExtension struct {
 	ExtensionPlus
 	HashLock         *HashLock
-	DstChainId       float32
+	DstChainId       int
 	DstToken         common.Address
 	SrcSafetyDeposit string
 	DstSafetyDeposit string
@@ -175,7 +175,7 @@ func decodeExtraData(data []byte) (*EscrowExtraData, error) {
 		HashLock: &HashLock{
 			hashlockData.String(),
 		},
-		DstChainId:       float32(dstChainIdData.Uint64()),
+		DstChainId:       int(dstChainIdData.Uint64()),
 		DstToken:         common.HexToAddress(addressHex),
 		SrcSafetyDeposit: srcSafetyDeposit,
 		DstSafetyDeposit: dstSafetyDeposit,
@@ -217,7 +217,7 @@ func decodeTimeLocks(value *big.Int) (*TimeLocks, error) {
 
 type EscrowExtraData struct {
 	HashLock         *HashLock
-	DstChainId       float32
+	DstChainId       int
 	DstToken         common.Address
 	SrcSafetyDeposit *big.Int
 	DstSafetyDeposit *big.Int
